Proyecto/requirements.txt
asn1crypto==1.2.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.8
cycler==0.10.0
decorator==4.4.1
feedparser==5.2.1
filelock==3.0.12
fpdf==1.7.2
funcsigs==1.0.2
html5lib==1.0.1
idna==2.8
importlib-metadata==0.23
inflection==0.3.1
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipython-sql==0.3.9
jedi==0.15.1
jupyter-client==5.3.4
jupyter-core==4.6.0
kiwisolver==1.1.0
langdetect==1.0.7
lxml==4.4.2
matplotlib==3.1.1
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
modin==0.6.3
more-itertools==5.0.0
ndg-httpsclient==0.5.1
numpy==1.17.2
packaging==19.2
pandas==0.25.3
parso==0.5.1
pathlib==1.0.1
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
pluggy==0.13.1
prettytable==0.7.2
prompt-toolkit==2.0.10
protobuf==3.11.0
psutil==5.6.7
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.8
pycparser==2.19
Pygments==2.4.2
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.5
PyPDF2==1.26.0
PySocks==1.7.1
pytest==5.3.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-resize-image==1.1.19
python-slugify==4.0.0
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.0
Quandl==3.4.8
ray==0.7.3
redis==3.3.11
requests==2.22.0
scipy==1.3.1
simple-imshow==1.0.post1
six==1.12.0
soupsieve==1.9.5
SQLAlchemy==1.3.10
sqlparse==0.3.0
statsmodels==0.10.1
stdiomask==0.0.4
text-unidecode==1.3
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
xlrd==1.2.0
zipp==0.6.0
